receipt-reconciliation-system/requirements.txt
# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database (MongoDB)
pymongo>=4.5.0
mongoengine>=0.27.0
motor>=3.3.0                    # Async MongoDB driver for better performance
#gridfs                          # MongoDB file storage (CRITICAL for PDF storage)

# PDF Processing
llama-index>=0.8.0
PyMuPDF>=1.24.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pdf2image                       # Convert PDF to images if OCR needed
Pillow>=10.0.0                 # Image processing support
pytesseract                    # OCR fallback for poor quality PDFs

# Text Processing & Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Data Validation & Utilities
pydantic>=2.0.0                # Data validation (CRITICAL for robustness)
python-dotenv>=1.0.0
python-dateutil>=2.8.2        # Better date parsing
requests>=2.31.0
tqdm>=4.65.0                  # Progress bars
python-multipart                # Streamlit file uploads
aiofiles                        # Async file operations
asyncio-throttle               # Rate limiting for email processing
tenacity>=8.2.0                # Retry mechanisms with exponential backoff

# Security
cryptography>=41.0.0           # Secure credential encryption
python-magic                    # File type validation (security requirement)
bleach                         # HTML sanitization
validators                     # Input validation

# Async Email Processing
aiosmtplib                     # Async SMTP support if needed
aioimaplib                     # Async IMAP (alternative to imaplib)

# Data Visualization (Free)
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
watchdog                        # File system monitoring

# Development
black>=23.0.0                  # Code formatting
flake8>=6.0.0                  # Linting
